pyproject.toml
[tool.poetry]
name = "web-scraping"
version = "0.1.0"
description = ""
authors = ["Adarsh047 <adarshshanmugam2002@gmail.com>"]
readme = "README.md"
packages = [{include = "web_scraping"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.10.0"
pillow = "^10.0.0"
webdriver-manager = "3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
